+++ core-service/src/main/resources/db/migration/V1__init.sql
@@ -15,7 +15,7 @@ create table products
     id          bigserial primary key,
     title       varchar(100),
     category_id bigint references categories (id),
-    price       int,
+    price       numeric(8, 2),
     created_at  timestamp default current_timestamp,
     updated_at  timestamp default current_timestamp
 );
@@ -42,8 +42,8 @@ values ('Milk', 70.00, 1),
 create table orders
 (
     id          bigserial primary key,
-    username    varchar(255) not null,
-    total_price int          not null,
+    username    varchar(255)  not null,
+    total_price numeric(8, 2) not null,
     address     varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -53,11 +53,11 @@ create table orders
 create table order_items
 (
     id                bigserial primary key,
-    product_id        bigint not null references products (id),
-    order_id          bigint not null references orders (id),
-    quantity          int    not null,
-    price_per_product int    not null,
-    price             int    not null,
+    product_id        bigint        not null references products (id),
+    order_id          bigint        not null references orders (id),
+    quantity          int           not null,
+    price_per_product numeric(8, 2) not null,
+    price             numeric(8, 2) not null,
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
 );
